+++ mlflow/java/src/main/java/org/mlflow/sagemaker/LeapFrameUtils.java
@@ -1,15 +1,15 @@
-package org.mlflow.mleap;
+package org.mlflow.mleap.sagemaker;
 
 import java.nio.charset.Charset;
 
 import ml.combust.mleap.json.DefaultFrameReader;
 import ml.combust.mleap.runtime.frame.DefaultLeapFrame;
 
-public class LeapFrameUtils {
+class LeapFrameUtils {
     private static final DefaultFrameReader frameReader = new DefaultFrameReader();
     private static final Charset jsonCharset = Charset.forName("UTF-8");
 
-    public static DefaultLeapFrame getLeapFrameFromJson(String frameJson) {
+    protected static DefaultLeapFrame getLeapFrameFromJson(String frameJson) {
         byte[] frameBytes = frameJson.getBytes(jsonCharset);
         return frameReader.fromBytes(frameBytes, jsonCharset).get();
     }
