+++ mlflow/java/src/test/resources/org/mlflow/mleap_model/sample_input.json
@@ -1,6 +1,7 @@
 [
     {
-        "topic": "alt.atheism", 
+        "fsfsdfsdfsf": "alt.atheism", 
+        "fsfsdfsdfsf7": "alt.atheism", 
         "text": "From: marshall@csugrad.cs.vt.edu (Kevin Marshall) Subject: Re: Faith and Dogma Organization: Virginia Tech Computer Science Dept, Blacksburg, VA Lines: 96 NNTP-Posting-Host: csugrad.cs.vt.edsadsadasdasdaddu   tgk@cs.toronto.edu (Todd Kelley) writes: >In light of what happened in Waco, I need to get something of my >chest. > >Faith and dogma are dangerous. "
     }
 ]
